Move helper methods for calendar to CalendarHelper module
