+++ app/controllers/calendar_controller.rb
@@ -1,5 +1,4 @@
 class CalendarController < ApplicationController
-  helper_method :get_event, :event_ongoing?, :event_span
 
   def index
     week_start = Date.new(2016, 6, 11)
@@ -19,25 +18,6 @@ class CalendarController < ApplicationController
       wday = event.start_time.wday
       @events[wday] << event
     end
-
-  end
-
-  def get_event(day, time)
-    datetime = add_date_and_time(day, time)
-    @events[day.wday].find do |event|
-      event.start_time == datetime
-    end
-  end
-
-  def event_ongoing?(day, time)
-    datetime = add_date_and_time(day, time)
-    @events[day.wday].any? do |event|
-      event.start_time < datetime && event.end_time > datetime
-    end
-  end
-
-  def event_span(event)
-    event ? (event.end_time - event.start_time) / 30.minutes : 1
   end
 
   private
@@ -45,7 +25,4 @@ class CalendarController < ApplicationController
     (start_time.to_i..end_time.to_i).step(step).collect { |time| Time.at time }
   end
 
-  def add_date_and_time(date, time)
-    date.to_datetime + time.seconds_since_midnight.seconds
-  end
 end
